Add a day of rework
I wasn't making any commits yesterday since it wasn't compiling. From now on, I'll make more commits even if the it can't compile, because it's becoming more important for me to be able to check older versions of the code.

This commit starts making an overhaul of the entire library, but it's not finished and it will probably not compile.