Parse errors with service version
The error handler can now detect if the error contains the service version and deserialize accordingly, identify which type of error it is and then have the error builder build the correct ApodError.

Before #16 is closed I have to parse the errors without service version and also write tests for the error handler.