Make StringBuilder local var instead of member
I was wrong about #18 having to do with objects not being disposed of correctly. The problem was that the ApodUriBuilder preserved the instance of the StringBuilder between requests, and since the URI was appended to the StringBuilder every time the URI was doubled which obviously did not work. This has been fixed by making the StringBuilder a local variable that is instantiated when the method is called.