Fix bug with fetching apods in date range
If the user set the startDate to DateTime.Today.AddDays(-3) and the endDate to DateTime.Today (ignoring the fact that it's Today by default) it *could* return an error if the user was ahead of Eastern Time and it was past local midnight but not past Eastern Time midnight. This commit fixes that.